Change default kill signal to INT
This is, potentially, a breaking change. However, my assumption is that most
apps run under ringleader (read: rails apps with script/server) are already
running in a PTY, so they think they're in a terminal, and thus INT is a better
(and more expected) signal.

Also, this is because of webrick, which oh-so-conveniently ignores TERM. When
that happens, even if it's being supervised by foreman, it leaves orphaned
processes around. Foreman would normally handle this situation, but it also dies
because its parent sub-shell (run by ringleader) handles TERM too, by closing,
and killing off foreman's STDOUT/STDERR. And foreman tries to log before killing
a process that didn't die gracefully, gets a SIGPIPE, and dies itself. And then
you end up with lonely little webrick, running as a child of 1...