Work around to fix #28.
There is an open issue (https://github.com/celluloid/celluloid-io/pull/23) in Celluloid for over two years and it is related with issue #28. @halorgium wrote a fix but it never got merged.

This commit is adding this fix as a 'monkey patch'. I know that it is not the best way to solve it but...

I asked to @wengzilla to test it and he reported me that it is working.